+++ R/tar_hook_outer.R
@@ -2,6 +2,9 @@
 #' @export
 #' @family hooks
 #' @description Wrap the command of each target in an arbitrary R expression.
+#'   `tar_hook_outer()` expects unevaluated expressions for the `hook` and
+#'   `names` arguments, whereas `tar_hook_outer_raw()` expects
+#'   evaluated expression objects.
 #' @details The expression you supply to `hook`
 #'   must contain the special placeholder symbol `.x`
 #'   so `tar_hook_outer()` knows where to insert the original command
@@ -16,8 +19,10 @@
 #'   The hook must contain the special placeholder symbol `.x`
 #'   so `tar_hook_outer()` knows where to insert the original command
 #'   of the target.
-#'   The hook code is quoted (not evaluated) so there is no need
-#'   to wrap it in `quote()`, `expression()`, or similar.
+#'
+#'   `tar_hook_outer()` expects unevaluated expressions for the `hook` and
+#'   `names` arguments, whereas `tar_hook_outer_raw()` expects
+#'   evaluated expression objects.
 #' @examples
 #' if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
 #' targets::tar_dir({ # tar_dir() runs code from a temporary directory.
@@ -38,30 +43,38 @@
 #'   )
 #' })
 #' targets::tar_manifest(fields = command)
+#' # Using tar_hook_outer_raw():
+#' targets::tar_script({
+#'   targets <- list(
+#'     # Nested target lists work with hooks.
+#'     list(
+#'       targets::tar_target(x1, task1()),
+#'       targets::tar_target(x2, task2(x1))
+#'     ),
+#'     targets::tar_target(x3, task3(x2)),
+#'     targets::tar_target(y1, task4(x3))
+#'   )
+#'   tarchetypes::tar_hook_outer_raw(
+#'     targets = targets,
+#'     hook = quote(postprocess(.x, arg = "value")),
+#'     names = quote(starts_with("x"))
+#'   )
+#' })
 #' })
 #' }
-tar_hook_outer <- function(targets, hook, names = NULL, set_deps = TRUE) {
-  targets::tar_assert_scalar(set_deps)
-  targets::tar_assert_lgl(set_deps)
-  targets::tar_assert_nonmissing(set_deps)
-  targets <- tar_copy_targets(targets)
-  hook <- substitute(hook)
-  targets::tar_assert_lang(hook)
-  assert_hook_placeholder(hook)
-  names_quosure <- rlang::enquo(names)
-  walk_targets(
+tar_hook_outer <- function(
+  targets,
+  hook,
+  names = NULL,
+  set_deps = TRUE,
+  envir = parent.frame()
+) {
+  force(envir)
+  tar_hook_outer_raw(
     targets = targets,
-    names_quosure = names_quosure,
-    fun = tar_hook_outer_insert,
-    hook = hook,
-    set_deps = set_deps
+    hook = substitute(hook),
+    names = substitute(names),
+    set_deps = set_deps,
+    envir = envir
   )
-  targets
-}
-
-tar_hook_outer_insert <- function(target, hook, set_deps) {
-  assert_hook_expr(target)
-  lang <- target$command$expr[[1]]
-  expr <- tar_sub_expr(hook, values = list(.x = lang))
-  tar_replace_command(target = target, expr = expr, set_deps = set_deps)
 }