+++ R/tar_map2_count.R
@@ -5,6 +5,12 @@
 #' @description Define targets for batched
 #'   dynamic-within-static branching for data frames,
 #'   where the user sets the (maximum) number of batches.
+#'
+#'   [tar_map2_count()] expects unevaluated language for arguments
+#'   `name`, `command1`, `command2`, `columns1`, and `columns2`.
+#'   [tar_map2_count_raw()] expects a character string for `name`
+#'   and an evaluated expression object  for each of
+#'   `command1`, `command2`, `columns1`, and `columns2`.
 #' @details Static branching creates one pair of targets
 #'   for each row in `values`. In each pair,
 #'   there is an upstream non-dynamic target that runs `command1`
@@ -18,7 +24,10 @@
 #' @inheritSection tar_rep Replicate-specific seeds
 #' @inheritParams tar_rep
 #' @inheritParams tar_map2
-#' @inheritParams tar_map2_count_raw
+#' @param batches Positive integer of length 1,
+#'   maximum number of batches (dynamic branches within static branches)
+#'   of the downstream (`command2`) targets. Batches
+#'   are formed from row groups of the `command1` target output.
 #' @examples
 #' if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
 #' targets::tar_dir({ # tar_dir() runs code from a temporary directory.
@@ -40,6 +49,27 @@
 #' })
 #' targets::tar_make()
 #' targets::tar_read(x)
+#' # With tar_map2_count_raw():
+#' targets::tar_script({
+#'   tarchetypes::tar_map2_count_raw(
+#'     name = "x",
+#'     command1 = quote(
+#'       tibble::tibble(
+#'         arg1 = arg1,
+#'         arg2 = seq_len(6)
+#'       )
+#'     ),
+#'     command2 = quote(
+#'       tibble::tibble(
+#'         result = paste(arg1, arg2),
+#'         random = sample.int(1e9, size = 1),
+#'         length_input = length(arg1)
+#'       )
+#'     ),
+#'     values = tibble::tibble(arg1 = letters[seq_len(2)]),
+#'     batches = 3
+#'    )
+#' })
 #' })
 #' }
 tar_map2_count <- function(
@@ -56,6 +86,7 @@ tar_map2_count <- function(
   columns1 = tidyselect::everything(),
   columns2 = tidyselect::everything(),
   rep_workers = 1,
+  delimiter = "_",
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
@@ -86,6 +117,7 @@ tar_map2_count <- function(
     columns1 = substitute(columns1),
     columns2 = substitute(columns2),
     rep_workers = rep_workers,
+    delimiter = delimiter,
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,